Simplified tread creation in main by removing lambdas. Change Philosopher::id type to size_t. Add info for not supported feature.

+++ Event.hpp
@@ -5,7 +5,7 @@
 enum class Action;
 
 struct Event {
-    int philosopher_id;
+    size_t philosopher_id;
     Action action;
     std::chrono::nanoseconds time;
     std::stringstream text{};
